Giving access of IgnoreHelper to the mods. Moving CommandReflectionHelper utils into ReflectionHelper. Removing static from IgnoreHelper.
