+++ SampleExternalMod/Injections/MainMenuHandler.cs
@@ -1,4 +1,5 @@
-﻿using HarmonyLib;
+﻿using CSM.Helpers;
+using HarmonyLib;
 using NLog;
 using SampleExternalMod.Commands;
 
@@ -13,10 +14,11 @@ namespace SampleExternalMod.Injections
         public static void Postfix(bool __result, ref byte district)
         {
             _logger.Info("Doing something");
-            CSMConnection.Instance.SentToAll(new TestCommand()
-            {
-                testing = "Heyyyy"
-            });
+            // CSMConnection.Instance.SentToAll(new TestCommand()
+            // {
+            //     testing = "Heyyyy"
+            // });
+            _logger.Info(IgnoreHelper.Instance.IsIgnored());
         }
     }
 }
