+++ src/CSM.csproj
@@ -237,13 +237,11 @@
     <Compile Include="Extensions\AreaExtension.cs" />
     <Compile Include="Extensions\DemandExtension.cs" />
     <Compile Include="Extensions\ThreadingExtension.cs" />
-    <Compile Include="Helpers\CommandReflectionHelper.cs" />
-    <Compile Include="Helpers\IgnoreHelper.cs" />
+    <Compile Include="Helpers\ReflectionHelper.cs" />
     <Compile Include="Helpers\SaveHelpers.cs" />
     <Compile Include="Helpers\ArrayXHelper.cs" />
     <Compile Include="Helpers\DLCHelper.cs" />
     <Compile Include="Helpers\InfoPanelHelper.cs" />
-    <Compile Include="Helpers\ReflectionHelper.cs" />
     <Compile Include="Helpers\SlowdownHelper.cs" />
     <Compile Include="Helpers\SpeedPauseHelper.cs" />
     <Compile Include="Helpers\ToolSimulator.cs" />
