--- src/Helpers/CommandReflectionHelper.cs
@@ -1,91 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Reflection;
-
-namespace CSM.Helpers
-{
-    public static class CommandReflectionHelper
-    {
-        // Not required, but prevents a number of spurious entries from making it to the log file.
-        private static readonly List<String> IgnoredAssemblies = new List<String>
-        {
-            "Anonymously Hosted DynamicMethods Assembly",
-            "Assembly-CSharp",
-            "Assembly-CSharp-firstpass",
-            "CSM.API",
-            "ColossalManaged",
-            "HarmonySharedState",
-            "ICSharpCode.SharpZipLib",
-            "ICities",
-            "LiteNetLib",
-            "mscorlib",
-            "Mono.Security",
-            "NLog",
-            "Open.Nat",
-            "protobuf-net",
-            "System.Threading",
-            "System",
-            "System.Configuration",
-            "System.Core",
-            "System.Xml",
-            "UnityEngine",
-            "UnityEngine.UI",
-            "UnityEngine.Networking",
-            "0Harmony",
-        };
-
-        /// <summary>
-        /// Searches all the assemblies in the current AppDomain for class definitions that implement the given class.
-        /// </summary>
-        public static IEnumerable<Type> FindClassesByType(Type typeToSearchFor)
-        {
-            var assemblies = AppDomain.CurrentDomain.GetAssemblies();
-
-            foreach (var assembly in assemblies)
-            {
-                var handlers = FetchHandlers(assembly, typeToSearchFor);
-                foreach (var handler in handlers)
-                {
-                    yield return handler;
-                }
-            }
-        }
-
-        private static IEnumerable<Type> FetchHandlers(Assembly assembly, Type typeToSearchFor)
-        {
-            var assemblyName = assembly.GetName().Name;
-
-            // Skip any assemblies that we don't anticipate finding anything in.
-            if (IgnoredAssemblies.Contains(assemblyName))
-            {
-                yield break;
-            }
-
-            Type[] types = new Type[0];
-            try
-            {
-                types = assembly.GetTypes();
-            }
-            catch
-            {
-            }
-
-            foreach (var type in types)
-            {
-                Boolean isValid = false;
-                try
-                {
-                    isValid = typeToSearchFor.IsAssignableFrom(type) && type.IsClass && !type.IsAbstract;
-                }
-                catch
-                {
-                }
-
-                if (isValid)
-                {
-                    yield return type;
-                }
-            }
-        }
-    }
-}
